cdc: use generation format v2 when introducing new generations
A node with this commit uses ONLY v2 when introducing new generations.

In particular, such a node will refuse to bootstrap into a cluster
that doesn't understand the v2 format (which is indicated by the
CDC_GENERATIONS_V2 feature).

The checkAndRepairCdcStreams command will also refuse to proceed,
asking the user to finish the upgrade first. If the user really wants
to run this command with the old format, they must run it on an old
node. Note that it doesn't reduce the "availability" of the command:
1. if all nodes have upgraded, the command will proceed on any node
2. if some node has not upgraded, the command will proceed on any of
   these old nodes.

New nodes DO understand the old format when learning
generations, they just refuse to introduce new generations using the
old format. Indeed: the entire purpose of the new format is to allow
larger generations to be created. In a later commit we shall remove the
generation truncating logic introduced in PR #8048.