sys_dist_ks: remove old internal CDC generations table definition
The old table won't be created in clusters that are bootstrapped after
this commit. It will stay in clusters that were upgraded
until the DB admin removes it.

Note that a fully upgraded cluster doesn't automatically create a new
generation in the new format. Even if the last generation was created
before the upgrade, the cluster will keep using it. Any new generation
will be created in the new format, and that happens when either:
1. a new node bootstraps,
2. or the user runs checkAndRepairCdcStreams, which has a new check: if
   the current generation uses the old format, the command will decide
   that repair is needed, even if the generation is completely fine
   otherwise.