dist/docker: return non-zero code when scylla setup fails
For the moment running Scylla with docker returns only 0 code
when process finishes it's execution.
It is suitable for the case of correctly setup Scylla and
shutdown/decomission operation, but not Scylla setup failure.

So, make containerized scylla return exit code "1" when
it's setup fails as in the following bug:
- https://github.com/scylladb/scylla/issues/8032

Such behavior is useful for proper handling of the Scylla containers
in such container orchestration systems as 'kubernetes'.
With current approach kubernetes doesn't know about suppressed error.
With the fix it will get to know that something goes wrong.