tracing: remove tracing::begin(lazy_eval)
To optimize code hot paths for tracing we are trying to remove dynamic
allocations and memory copying as much as possible. One of the biggest
contributor to the performance overhead in tracing is string formatting.
There was only 1 usage of this form of begin and the only reason it
existed was to delay argument formatting until the implementation
checked that the trace_state is initialized. For the fast slow query
mode we are trying to omit doing dyn allocs and copies of the values.

This patch gracefully sacrificies the flexibility of this call in favor
of better performance.

Signed-off-by: Ivan Prisyazhnyy <ivan@scylladb.com>