tracing: avoid copying a query string in a prepared statements tracing hot path
To optimize code hot paths for tracing we are trying to remove dynamic
allocations and memory copying as much as possible. One of the biggest
contributor to the performance overhead in tracing is string formatting.

The hottest code path is expected to be the execution of prepared
statements. A prepared statement usually has a single query to trace.
Usually prep statements are all compiled and cached. Thus we can omit
copying and allocating memory of keeping the track of the related query.

This patch changes the way trace state params tracks queries for cached
prepared statement by reusing prepared_checked_weak_ptr. This allows to save 1 dyn
allocation and copy on the hot path.

Signed-off-by: Ivan Prisyazhnyy <ivan@scylladb.com>