commitlog: Fix inner loop condition in allocation pre-fill
Fixes #8369

This was originally found (and fixed) by @gleb-cloudius, but the patch set with the
fix was reverted at some point, and the fix went away. Now the error remains even
in new, nice coroutine code.

We check the wrong var in the inner loop of the pre-fill path of allocate_segment_ex,
often causing us to generate giant writev:s of more or less the whole file.
Not intended.